+++ app/src/main/java/com/songlib/app/navigation/AppNavHost.kt
@@ -25,8 +25,8 @@ import com.songlib.core.database.model.SongEntity
 import com.songlib.feature.donation.DonationViewModel
 import com.songlib.feature.donation.view.DonationScreen
 import com.songlib.feature.donation.view.PaymentWebViewScreen
-import com.songlib.feature.broadcast.BroadcastViewModel
-import com.songlib.feature.broadcast.view.BroadcastScreen
+import com.songlib.feature.casting.CastingViewModel
+import com.songlib.feature.casting.view.CastingScreen
 import com.songlib.feature.edits.admin.AdminEditsViewModel
 import com.songlib.feature.edits.admin.view.AdminEditsScreen
 import com.songlib.feature.drafts.list.DraftsViewModel
@@ -265,9 +265,9 @@ fun AppNavHost(
             AdminEditsScreen(navController = navController, viewModel = viewModel)
         }
 
-        composable(Routes.BROADCAST) {
-            val viewModel: BroadcastViewModel = hiltViewModel()
-            BroadcastScreen(navController = navController, viewModel = viewModel)
+        composable(Routes.CASTING) {
+            val viewModel: CastingViewModel = hiltViewModel()
+            CastingScreen(navController = navController, viewModel = viewModel)
         }
     }
 }
