--- core/broadcast/src/main/AndroidManifest.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <!-- Lets the server bind a socket and lets us read the device's local IP(s) -->
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
-
-    <!-- Needed to keep the embedded server alive in a foreground service -->
-    <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
-    <uses-permission android:name="android.permission.FOREGROUND_SERVICE_CONNECTED_DEVICE" />
-    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
-
-    <application>
-        <service
-            android:name="com.songlib.core.broadcast.service.BroadcastForegroundService"
-            android:exported="false"
-            android:foregroundServiceType="connectedDevice" />
-    </application>
-
-</manifest>
