+++ feature/drafts/src/main/java/com/songlib/feature/drafts/present/DraftPresenterViewModel.kt
@@ -2,7 +2,7 @@ package com.songlib.feature.drafts.present
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.songlib.core.broadcast.PresentationBroadcastRepo
+import com.songlib.core.casting.CastingRepo
 import com.songlib.core.common.entity.UiState
 import com.songlib.core.common.utils.getSongVerses
 import com.songlib.core.data.repos.DraftRepo
@@ -20,7 +20,7 @@ import javax.inject.Inject
 @HiltViewModel
 class DraftPresenterViewModel @Inject constructor(
     private val draftRepo: DraftRepo,
-    private val broadcastRepo: PresentationBroadcastRepo,
+    private val broadcastRepo: CastingRepo,
 ) : ViewModel() {
 
     private val _uiState = MutableStateFlow<UiState>(UiState.Loading)
