+++ feature/home/src/main/java/com/songlib/feature/home/components/HomeTopBar.kt
@@ -66,8 +66,8 @@ fun HomeAppBar(
                         }
                     }
 
-                    IconButton(onClick = { navController.navigate(Routes.BROADCAST) }) {
-                        Icon(Icons.Default.Cast, contentDescription = "Broadcast to PC")
+                    IconButton(onClick = { navController.navigate(Routes.CASTING) }) {
+                        Icon(Icons.Default.Cast, contentDescription = "Casting to PC")
                     }
 
                     if (hasHistory) {
