+++ feature/song/build.gradle.kts
@@ -10,6 +10,6 @@ android {
 dependencies {
     implementation(project(":core:common"))
     implementation(project(":core:data"))
-    implementation(project(":core:broadcast"))
+    implementation(project(":core:casting"))
     implementation(project(":feature:home"))
 }
